+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2019-07-23
+### Added
+- Added the ability to specify whether or not the application name is emitted as part of output using `--emit-app-name`.
+- Added the ability to notify only on failures using `--notify-only-on-failure`.
+- Added the ability to provide more verbose output only where a failure occurs to help diagnose the issue using --verbose-on-failure`.
+
 ## [1.0.0] - 2019-07-11
 ### Added
 - Return code emitted to fail CI jobs where upload fails.